supabase client class & push token helper
